Add Slack integration
* Objective

This commit objective is about adding the Slack integration that will be
used to receive and process the incoming messages using a callback
function.

* Why

The main purpose of the project is the Slack integration we need the
functionality of receiving, processing, and send the results back to
Slack.

* How

This commit introduces the Slack integration using the Slack RTM API.